+++ changelog.md
@@ -1,3 +1,8 @@
+1.1.0
+=================================
+- Cleanup material-ui
+- Remove schema generator
+
 1.0.6
 =================================
 - Public release